\name{dfdb.RODBC-package}
\alias{dfdb.RODBC-package}
\alias{dfdb.RODBC}
\docType{package}
\title{
Access Data Tables in an RODBC Data Base as Pseudo-Dataframes
}
\description{
This package contains functions that allow the user to treat data
  tables in an ODBC-accessible data base as if they were data frames, including
  the ability to access subsets of the rows and columns of the table and to
  modify the table; row "subscripts" can be specified as SQL queries.
}
\details{
\tabular{ll}{
Package: \tab dfdb.RODBC\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2008-11-29\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
John Fox and Oswaldo Cruz

Maintainer: John Fox <jfox@mcmaster.ca>
}

\keyword{ package }
\keyword{ manip }

\seealso{
\code{\link[RODBC]{RODBC}}
}
